+++ passenger-service/src/test/java/com/modsen/passengerservice/integration/kafka/KafkaConsumerIT.java
@@ -1,6 +1,7 @@
 package com.modsen.passengerservice.integration.kafka;
 
 import com.modsen.passengerservice.dto.message.PassengerRatingMessage;
+import com.modsen.passengerservice.integration.TestcontainersBase;
 import com.modsen.passengerservice.repository.PassengerRepository;
 import com.modsen.passengerservice.service.TestSender;
 import lombok.RequiredArgsConstructor;
@@ -8,14 +9,7 @@ import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.DynamicPropertyRegistry;
-import org.springframework.test.context.DynamicPropertySource;
 import org.springframework.test.context.TestConstructor;
-import org.testcontainers.containers.KafkaContainer;
-import org.testcontainers.containers.PostgreSQLContainer;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
-import org.testcontainers.utility.DockerImageName;
 
 import java.time.Duration;
 
@@ -25,36 +19,12 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.awaitility.Awaitility.await;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@Testcontainers
 @TestConstructor(autowireMode = TestConstructor.AutowireMode.ALL)
 @RequiredArgsConstructor
-public class KafkaConsumerIT {
+public class KafkaConsumerIT extends TestcontainersBase {
 
     private final PassengerRepository passengerRepository;
     private final TestSender testSender;
-
-    @Container
-    private static final PostgreSQLContainer<?> postgres = new PostgreSQLContainer<>(
-            "postgres:15-alpine"
-    );
-
-    @Container
-    private static final KafkaContainer kafka = new KafkaContainer(
-            DockerImageName.parse("confluentinc/cp-kafka:7.3.3")
-    );
-
-    @DynamicPropertySource
-    private static void postgresProperties(DynamicPropertyRegistry registry) {
-        registry.add("spring.datasource.url", postgres::getJdbcUrl);
-        registry.add("spring.datasource.username", postgres::getUsername);
-        registry.add("spring.datasource.password", postgres::getPassword);
-        registry.add("postgresql.driver", postgres::getDriverClassName);
-    }
-
-    @DynamicPropertySource
-    private static void kafkaProperties(DynamicPropertyRegistry registry) {
-        registry.add("spring.kafka.bootstrap-servers", kafka::getBootstrapServers);
-    }
 
     @BeforeAll
     static void beforeAll() {
